+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.0] - TBD
+
+### Fixed
+
+### Changed
+
+- Merged inbound and outbound routes to declutter the map.
+- Route paths are now great circle paths rather than bezier curves.
+- Route popup is now tabbed to seperate inbound and outbound pairings.
+- Pairings are now listed alphabetically.
+
+### Added
+
+- Mouse-over increases width of route paths.
+- Mouse-over increases size of destination markers.
+- Selective data download. Ability to choose which months to download. Size of data is also shown prior to download. Warning for more than 50mb on non-desktop devices.
+- Depature count added to destination marker popup.
+- Overnight tab added to destination marker popup. Disabled if there are no overnights.
+
 ## [1.1.7] - 2024-12-14
 
 ### Fixed